+++ src/main/java/com/zb/deuggeun/program/repository/ProgramRepository.java
@@ -1,6 +1,6 @@
 package com.zb.deuggeun.program.repository;
 
-import com.zb.deuggeun.common.service.CustomJpaRepository;
+import com.zb.deuggeun.common.repository.CustomJpaRepository;
 import com.zb.deuggeun.member.entity.Member;
 import com.zb.deuggeun.program.entity.Program;
 import com.zb.deuggeun.program.type.ProgramStatus;
